+++ common/src/main/java/net/creeperhost/equivalentexchange/compat/teams/TeamsKnowledgeHandler.java
@@ -26,13 +26,7 @@ public class TeamsKnowledgeHandler implements IKnowledgeHandler
     @Override
     public Path getSavePath(Player player)
     {
-        return getBaseSavePath().resolve(getPlayerTeam(player).getTeamId().toString() + "_knowledge.dat");
-    }
-
-    @Override
-    public Path getBaseSavePath()
-    {
-        return FTBTeamsAPI.api().getManager().getServer().getWorldPath(LevelResource.ROOT).resolve("ftbteams/knowledge/");
+        return FTBTeamsAPI.api().getManager().getServer().getWorldPath(LevelResource.ROOT).resolve("ftbteams/knowledge/").resolve(getPlayerTeam(player).getTeamId().toString() + "_knowledge.dat");
     }
 
     private Team getPlayerTeam(Player player)
@@ -125,7 +119,7 @@ public class TeamsKnowledgeHandler implements IKnowledgeHandler
     public void removeKnowledge(Player player, ItemStack stack)
     {
         List<ItemStack> stacks = getKnowledgeList(player);
-        stacks.remove(stack);
+        stacks.removeIf(itemStack -> itemStack.is(stack.getItem()));
         setKnowledgeList(player, stacks);
         if(!player.level().isClientSide)
             KnowledgeChangedEvent.KNOWLEDGE_REMOVED_EVENT.invoker().removed(player, stack);
@@ -135,7 +129,7 @@ public class TeamsKnowledgeHandler implements IKnowledgeHandler
     public void removeKnowledge(UUID uuid, ItemStack stack)
     {
         List<ItemStack> stacks = getKnowledgeList(uuid);
-        stacks.remove(stack);
+        stacks.removeIf(itemStack -> itemStack.is(stack.getItem()));
         setKnowledgeList(uuid, stacks);
     }
 
@@ -144,6 +138,8 @@ public class TeamsKnowledgeHandler implements IKnowledgeHandler
     {
         if(player != null && !player.level().isClientSide)
         {
+            getSavePath(player).getParent().toFile().mkdirs();
+
             EquivalentExchange.LOGGER.info("Saving knowledge for team " + getPlayerTeam(player).getTeamId().toString());
             try
             {