+++ common/src/main/java/net/creeperhost/equivalentexchange/inventory/TransmutationInventory.java
@@ -4,9 +4,11 @@ import net.creeperhost.equivalentexchange.EquivalentExchange;
 import net.creeperhost.equivalentexchange.api.EquivalentExchangeAPI;
 import net.creeperhost.equivalentexchange.api.item.IKleinStarItem;
 import net.creeperhost.equivalentexchange.impl.TransmutationTableHandler;
+import net.creeperhost.equivalentexchange.network.packets.knowledge.RemoveKnowledgePacket;
 import net.minecraft.core.NonNullList;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.nbt.ListTag;
+import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.Container;
 import net.minecraft.world.ContainerHelper;
 import net.minecraft.world.entity.player.Player;
@@ -47,7 +49,7 @@ public class TransmutationInventory implements Container
     @Override
     public @NotNull ItemStack getItem(int i)
     {
-        return i >= 0 && i < this.items.size() ? this.items.get(i) : ItemStack.EMPTY;
+        return  items.get(i);
     }
 
     @Override
@@ -61,7 +63,7 @@ public class TransmutationInventory implements Container
                 EquivalentExchangeAPI.getStorageHandler().removeEmcFor(player, EquivalentExchangeAPI.getEmcValue(itemStack) * j);
             }
         }
-        if (!itemStack.isEmpty()) this.setChanged();
+        if (!itemStack.isEmpty()) this.updateInventory();
         return itemStack;
     }
 
@@ -79,60 +81,71 @@ public class TransmutationInventory implements Container
     public void setItem(int i, @NotNull ItemStack itemStack)
     {
         if(player == null) return;
-        if(!EquivalentExchangeAPI.hasEmcValue(itemStack))
+        boolean client = player.level().isClientSide();
+
+        if(!client && !EquivalentExchangeAPI.hasEmcValue(itemStack))
         {
             EquivalentExchange.LOGGER.error("Player tried adding item {} to transmutation inventory with no emc value", itemStack);
             return;
         }
-        //Burn slot
-        if(i == 0)
+        if(i != 0)
+            this.items.set(i, itemStack);
+        if(!client)
         {
-            if(!getItem(i).isEmpty())
+            // SLOT: Burn item
+            if(i == 0)
             {
-                if(!getItem(i).is(Items.AIR))
+                if(!getItem(i).isEmpty())
+                {
                     EquivalentExchange.LOGGER.error("Item stuck in burn slot {} with emc value {}", getItem(i).getDisplayName().getString(), EquivalentExchangeAPI.getEmcValue(itemStack));
-            }
-            //Create a new instance of the item in order to remove any extra data added in other ways
-            EquivalentExchangeAPI.getKnowledgeHandler().addKnowledge(player, new ItemStack(itemStack.getItem()));
+                }
+                //Create a new instance of the item in order to remove any extra data added in other ways
+                EquivalentExchangeAPI.getKnowledgeHandler().addKnowledge(player, new ItemStack(itemStack.getItem()));
 
-            double value = EquivalentExchangeAPI.getEmcValue(itemStack) * itemStack.getCount();
-            if(itemStack.getItem() instanceof IKleinStarItem kleinStar)
-            {
-                value += kleinStar.getKleinStarStored(itemStack);
+                double value = EquivalentExchangeAPI.getEmcValue(itemStack) * itemStack.getCount();
+                if(itemStack.getItem() instanceof IKleinStarItem kleinStar)
+                {
+                    value += kleinStar.getKleinStarStored(itemStack);
+                }
+                EquivalentExchangeAPI.getStorageHandler().addEmcFor(player, value);
             }
-            EquivalentExchangeAPI.getStorageHandler().addEmcFor(player, value);
+            //SLOT: Unlearn item
+            if(i == 1)
+            {
+                EquivalentExchangeAPI.getKnowledgeHandler().removeKnowledge(player, itemStack);
+                new RemoveKnowledgePacket(itemStack).sendTo((ServerPlayer) player);
+            }
+            //SLOTS: Center of the right side of tablet (2); left side of the tablet ([3..10])
+            else if(i >= 2 && i <= 10)
+            {
+                if(itemStack.getItem() instanceof IKleinStarItem kleinStar)
+                {
+                    double starCurrent = kleinStar.getKleinStarStored(itemStack);
+                    double playerCurrent = EquivalentExchangeAPI.getStorageHandler().getEmcValueFor(player);
+                    //Fill klein star
+                    if(i != 2 && starCurrent < kleinStar.getKleinStarMaxStorage(itemStack))
+                    {
+                        double starMax = kleinStar.getKleinStarMaxStorage(itemStack);
+                        double starFree = starMax - starCurrent;
+                        double starAdded = Math.min(playerCurrent, starFree);
+
+                        kleinStar.setKleinStarEmc(itemStack, starCurrent + starAdded);
+                        EquivalentExchangeAPI.getStorageHandler().removeEmcFor(player, starAdded);
+                    }
+                    //Drain klein star
+                    else if(i == 2 && starCurrent > 0)
+                    {
+                        kleinStar.setKleinStarEmc(itemStack, 0);
+                        EquivalentExchangeAPI.getStorageHandler().addEmcFor(player, starCurrent);
+                    }
+                }
+                if(EquivalentExchangeAPI.hasEmcValue(itemStack))
+                {
+                    EquivalentExchangeAPI.getKnowledgeHandler().addKnowledge(player, new ItemStack(itemStack.getItem()));
+                }
+            }
             updateInventory();
-            //return here so the item is not added to the items list
-            return;
         }
-        else if(i == 1)
-        {
-            EquivalentExchangeAPI.getKnowledgeHandler().removeKnowledge(player, getItem(i));
-            updateInventory();
-        }
-        else if(i >= 3 && i <= 10)
-        {
-            if(itemStack.getItem() instanceof IKleinStarItem kleinStar && kleinStar.getKleinStarStored(itemStack) < kleinStar.getKleinStarMaxStorage(itemStack))
-            {
-                double starCurrent = kleinStar.getKleinStarStored(itemStack);
-                double starMax = kleinStar.getKleinStarMaxStorage(itemStack);
-                double playerCurrent = EquivalentExchangeAPI.getStorageHandler().getEmcValueFor(player);
-                double needed = starMax - starCurrent;
-
-                double inserted = Math.min(playerCurrent, needed);
-
-                kleinStar.setKleinStarEmc(itemStack, inserted);
-                double value = EquivalentExchangeAPI.getStorageHandler().getEmcValueFor(player) - inserted;
-                EquivalentExchangeAPI.getStorageHandler().setEmcValueFor(player, value);
-                updateInventory();
-            }
-            else if(EquivalentExchangeAPI.hasEmcValue(itemStack))
-            {
-                EquivalentExchangeAPI.getKnowledgeHandler().addKnowledge(player, itemStack);
-                updateInventory();
-            }
-        }
-        this.items.set(i, itemStack);
         setChanged();
     }
 
@@ -145,7 +158,7 @@ public class TransmutationInventory implements Container
     {
         if(player == null) return;
 
-        NonNullList<ItemStack> transmutations = TransmutationTableHandler.getTransmutationContent(page, filter, player);
+        NonNullList<ItemStack> transmutations = TransmutationTableHandler.getTransmutationContent(page, filter, getItem(2), player);
         int start = 11;
         int end = 26;
         int i = start;