+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.9
+
+### Fixed
+- Fixed Transmutation inventory not syncing correctly due to code being called on the client
+- Fixed hard dep on FTB Teams
 ## 1.0.8
 
 ### Fixed